requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2==2.9.7  # PostgreSQL adapter
django-cors-headers==4.2.0
djangorestframework-simplejwt==5.3.0  # JWT authentication
drf-yasg==1.21.4

